Fix config of ElasticSearch host/port in app.conf
