+++ conf/application.conf
@@ -303,6 +303,9 @@ swagger.api.info {
 
 env.local = ${env.default} {
 
+  elasticsearch.host = "localhost"
+  elasticsearch.port = 9200
+
   elastic.recreate.index = true
   elastic.import.sample = true
   elasticsearch.bi.name = "bi-dev"
@@ -315,6 +318,9 @@ env.local = ${env.default} {
 
 env.default {
   # duplicated config ...
+  elasticsearch.host = ${?ONS_BI_API_ES_HOST}
+  elasticsearch.port = ${?ONS_BI_API_ES_PORT}
+
   elastic.recreate.index = ${?ONS_BI_API_CREATE_INDEX}
   elastic.import.sample = false
 
@@ -337,6 +343,9 @@ env.default {
 }
 
 env.test: ${env.local} {
+  elasticsearch.host = "localhost"
+  elasticsearch.port = 9200
+
   elasticsearch.local = true
   elasticsearch.uri = "elasticsearch://localhost:9200"
 }
